fet-example/requirements.txt
uagents==0.23.6
uagents-core==0.4.0
python-dotenv
google-genai
Pillow